Add AmmoManager which handles consumption and adding of ammo

+++ AstroGod/Assets/Scripts/Items/WeaponData.cs
@@ -9,7 +9,7 @@ public class WeaponData : ItemData
     [field : SerializeField] public float BaseFireRate { get; private set; }
     [field : SerializeField] public float BaseDamage { get; private set; }
     [field: SerializeField] public float FirePower { get; private set; } // Affects the force at which projectiles are shot
-    [field: SerializeField] public float AmmoCost { get; private set; }
+    [field: SerializeField] public int AmmoCost { get; private set; }
     [field: SerializeField] public ProjectileController ProjectilePrefab { get; private set; }
 
 }
